4-reasoning-engine/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-openai
langchain-core
pydantic
python-dotenv
